+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Missing markdown file rename support from marksman LSP
   - [Github issue](https://github.com/artempyanykh/marksman/issues/153)
 
-## WIP
+## v2.9.2
 
 - feat(statix): Install statix from Mason
 - feat(solarize): Override colors on matching parentheses